+++ packages/akashic-cli-serve/src/client/akashic/AkashicGameViewWeb.d.ts
@@ -40,6 +40,7 @@ declare module agv {
 		};
 		playConfig: PlaylogConfig;
 		gameLoaderCustomizer: GameLoaderCustomizer;
+		argument?: any;
 	}
 	interface ErrorListener {
 		onError: (e: Error) => void;