+++ packages/akashic-cli-serve/src/client/store/Store.ts
@@ -7,11 +7,13 @@ import {LocalInstanceEntity} from "./LocalInstanceEntity";
 import {DevtoolUiStore} from "./DevtoolUiStore";
 import {ToolBarUiStore} from "./ToolBarUiStore";
 import {storage} from "./storage";
+import {StartupScreenUiStore} from "./StartupScreenUiStore";
 
 export class Store {
 	@observable playStore: PlayStore;
 	@observable toolBarUiStore: ToolBarUiStore;
 	@observable devtoolUiStore: DevtoolUiStore;
+	@observable startupScreenUiStore: StartupScreenUiStore;
 	@observable player: Player | null;
 
 	@observable currentPlay: PlayEntity | null;
@@ -22,6 +24,7 @@ export class Store {
 		this.playStore = new PlayStore();
 		this.toolBarUiStore = new ToolBarUiStore();
 		this.devtoolUiStore = new DevtoolUiStore();
+		this.startupScreenUiStore = new StartupScreenUiStore();
 		this.player = { id: storage.data.playerId, name: storage.data.playerName };
 		this.currentPlay = null;
 		this.currentLocalInstance = null;