+++ packages/akashic-cli-serve/src/server/route/ConfigRoute.ts
@@ -1,13 +1,18 @@
 import * as express from "express";
-import { createHandlerToGetEngineConfig } from "../controller/ConfigController";
+import { createHandlerToGetEngineConfig, handleToGetStartupOptions } from "../controller/ConfigController";
 
-export const createConfigRouter = (): express.Router => {
+export interface ConfigRouterParameterObject {
+	targetDir: string;
+}
+
+export const createConfigRouter = (params: ConfigRouterParameterObject): express.Router => {
 	const configRouter = express.Router();
 
-	configRouter.get("/content.json", createHandlerToGetEngineConfig(false));
+	configRouter.get("/content.json", createHandlerToGetEngineConfig(params.targetDir, false));
 	// /engineとの相違点はスクリプトアセット加工前のコンテンツを含む情報を投げること
 	// サーバー側でインスタンスを立ち上げる時は加工前のスクリプトアセットを参照する必要がある
-	configRouter.get("/content.raw.json", createHandlerToGetEngineConfig(true));
+	configRouter.get("/content.raw.json", createHandlerToGetEngineConfig(params.targetDir, true));
+	configRouter.get("/options", handleToGetStartupOptions);
 
 	return configRouter;
 };